+++ kotlin-stdlib/commonTest/kotlin/com/javiersc/kotlin/stdlib/CollectionsTest.kt
@@ -2,8 +2,8 @@
 
 package com.javiersc.kotlin.stdlib
 
+import io.kotest.matchers.shouldBe
 import kotlin.test.Test
-import kotlin.test.assertTrue
 
 class CollectionsTest {
 
@@ -11,26 +11,26 @@ class CollectionsTest {
 
     @Test
     fun collection_second() {
-        assertTrue { numbers.second() == 2 }
+        numbers.second().shouldBe(2)
     }
 
     @Test
     fun collection_third() {
-        assertTrue { numbers.third() == 3 }
+        numbers.third().shouldBe(3)
     }
 
     @Test
     fun collection_forth() {
-        assertTrue { numbers.forth() == 4 }
+        numbers.forth().shouldBe(4)
     }
 
     @Test
     fun collection_fifth() {
-        assertTrue { numbers.fifth() == 5 }
+        numbers.fifth().shouldBe(5)
     }
 
     @Test
     fun collection_penultimate() {
-        assertTrue { numbers.penultimate() == 9 }
+        numbers.penultimate().shouldBe(9)
     }
 }