+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 ### Fixed
 
 ### Updated
+- `org.jetbrains.kotlin:kotlin-gradle-plugin -> 1.6.0`
 
 
 ## [0.1.0-alpha.1] - 2021-11-10