+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 ### Fixed
 
 ### Updated
+- `com.javiersc.gradle-plugins:all-plugins -> 0.1.0-rc.11`
 
 
 ## [0.1.0-alpha.1] - 2021-11-10