+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 ### Fixed
 
 ### Updated
+- `org.jetbrains.kotlin:kotlin-gradle-plugin -> 1.6.0`
 - `com.javiersc.gradle-plugins:all-plugins -> 0.1.0-rc.11`