+++ CHANGELOG.md
@@ -16,8 +16,8 @@
 ### Fixed
 
 ### Updated
+- `com.javiersc.gradle-plugins:all-plugins -> 0.1.0-rc.12`
 - `org.jetbrains.kotlin:kotlin-gradle-plugin -> 1.6.0`
-- `com.javiersc.gradle-plugins:all-plugins -> 0.1.0-rc.11`
 
 
 ## [0.1.0-alpha.1] - 2021-11-10