+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 
 ### Added
+- `emptyString()`
 - `isNotNullNorBlank()` and `isNotNullNorEmpty()`
 - `second()`, `third()`, `forth()`, `fifth()` and `penultimate()`
 - `OrNull` variants for `second()`, `third()`, `forth()`, `fifth()` and `penultimate()`