+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 ### Fixed
 
 ### Updated
+- `io.kotest:kotest-assertions-core -> 5.0.0`
 - `com.javiersc.gradle-plugins:all-plugins -> 0.1.0-rc.12`
 - `org.jetbrains.kotlin:kotlin-gradle-plugin -> 1.6.0`