+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 
 ### Added
+
 - `emptyString()`
 - `isNotNullNorBlank()` and `isNotNullNorEmpty()`
 - `second()`, `third()`, `forth()`, `fifth()` and `penultimate()`
@@ -17,21 +18,24 @@
 ### Fixed
 
 ### Updated
-- `io.kotest:kotest-assertions-core -> 5.0.0`
-- `com.javiersc.gradle-plugins:all-plugins -> 0.1.0-rc.12`
-- `org.jetbrains.kotlin:kotlin-gradle-plugin -> 1.6.0`
 
+- `com.javiersc.gradle-plugins:all-plugins -> 0.1.0-rc.17`
+- `io.kotest:kotest-assertions-core -> 5.0.0`
+- `org.jetbrains.kotlin:kotlin-gradle-plugin -> 1.6.0`
 
 ## [0.1.0-alpha.1] - 2021-11-10
 
 ### Added
+
 - `AnsiColor` and `String.ansiColor()`
 - `String.remove`
 
 ### Updated
+
 - `gradle -> 7.3`
 - `com.javiersc.gradle-plugins:all-plugins -> 0.1.0-rc.8`
 - `org.jetbrains.kotlinx:binary-compatibility-validator -> 0.8.0`
 
 ## [0.0.0] - 2021-11-04
+
 - No changes