+++ CHANGELOG.md
@@ -14,6 +14,8 @@
 
 ### Updated
 
+- `com.javiersc.gradle-plugins:all-plugins -> 0.1.0-rc.22`
+
 ## [0.1.0-alpha.3] - 2021-11-29
 
 ### Changed