+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 
 ### Updated
 
+- `io.kotest:kotest-assertions-core -> 5.0.1`
 - `com.javiersc.gradle-plugins:all-plugins -> 0.1.0-rc.22`
 
 ## [0.1.0-alpha.3] - 2021-11-29