+++ CHANGELOG.md
@@ -14,9 +14,9 @@
 
 ### Updated
 
+- `com.javiersc.gradle-plugins:all-plugins -> 0.1.0-rc.24`
 - `io.kotest:kotest-assertions-core -> 5.0.2`
 - `gradle -> 7.3.1`
-- `com.javiersc.gradle-plugins:all-plugins -> 0.1.0-rc.22`
 
 ## [0.1.0-alpha.3] - 2021-11-29