+++ CHANGELOG.md
@@ -14,11 +14,11 @@
 
 ### Updated
 
+- `com.javiersc.gradle-plugins:all-plugins -> 0.1.0-rc.39`
 - `org.jetbrains.kotlinx:kotlinx-coroutines-core -> 1.6.0`
 - `gradle -> 7.3.3`
 - `io.kotest:kotest-assertions-core -> 5.0.3`
 - `org.jetbrains.kotlin:kotlin-gradle-plugin -> 1.6.10`
-- `com.javiersc.gradle-plugins:all-plugins -> 0.1.0-rc.24`
 
 ## [0.1.0-alpha.3] - 2021-11-29